requirements.txt
# Yahoo Fantasy API
yahoo_fantasy_api>=2.5.0
yahoo_oauth>=2.0.0

# HTTP requests
requests>=2.28.0

# Environment management
python-dotenv>=1.0.0

# Data sources (optional - for advanced stats)
# pybaseball>=2.2.0

# Scheduling (for local runs)
# schedule>=1.2.0
